refactor/extract-use-search-product-collection-hook
